Рефакторинг и устранение замечаний по отрисовке пунктов предложений
