+++ src/server/points-api-server.js
@@ -1,4 +1,5 @@
 import ApiService from '../framework/api-service.js';
+import AdapterService from './adapter.js';
 
 const Method = {
   GET: 'GET',
@@ -6,6 +7,8 @@ const Method = {
 };
 
 export default class PointsApiService extends ApiService {
+  #pointsAdapterService = new AdapterService();
+
   get points() {
     return this._load({url: 'points'})
       .then(ApiService.parseResponse);
@@ -21,32 +24,16 @@ export default class PointsApiService extends ApiService {
       .then(ApiService.parseResponse);
   }
 
-  async updateTask(point) {
+  async updatePoint(point) {
     const response = await this._load({
       url: `points/${point.id}`,
       method: Method.PUT,
-      body: JSON.stringify(this.adaptToServer(point)),
+      body: JSON.stringify(this.#pointsAdapterService.adaptToServer(point)),
       headers: new Headers({'Content-Type': 'application/json'}),
     });
 
     const parsedResponse = await ApiService.parseResponse(response);
 
     return parsedResponse;
-  }
-
-  #adaptToServer(point) {
-    const adaptedPoint = {...point,
-      'base_price': point.basePrice,
-      'date_from': point.dateFrom instanceof Date ? point.dateFrom.toISOString() : null,
-      'date_to': point.dateTo instanceof Date ? point.dateTo.toISOString() : null,
-      'is_favorite': point.isFavorite,
-    };
-
-    delete adaptedPoint.basePrice;
-    delete adaptedPoint.dateFrom;
-    delete adaptedPoint.dateTo;
-    delete adaptedPoint.isFavorite;
-
-    return adaptedPoint;
   }
 }
