+++ src/view/editor-form-elements.js
@@ -1,5 +1,5 @@
 
-const createOfferItemTemplate = (type, title, price, className, id) => `
+const createOfferItemTemplate = (type, title, price, id, className) => `
   <div class="event__offer-selector">
     <input class="event__offer-checkbox  visually-hidden" id="event-offer-${type}-${id}" type="checkbox" name="event-offer-${type}" data-offer-id="${id}" ${className}>
     <label class="event__offer-label" for="event-offer-${type}-${id}">
